+++ build-config/fragments/dev.js
@@ -22,7 +22,7 @@ module.exports = {
         // webpack-dev-server middleware
         before(app) {
             // use proper mime-type for wasm files
-            app.get('*.wasm', function(req, res, next) {
+            app.get('*.wasm', (req, res, next) => {
                 var options = {
                     root: contentBase,
                     dotfiles: 'deny',
@@ -30,14 +30,14 @@ module.exports = {
                         'Content-Type': 'application/wasm'
                     }
                 };
-                res.sendFile(req.url, options, function (err) {
+                res.sendFile(req.url, options, (err) => {
                     if (err) {
                         next(err);
                     }
                 });
             });
             // file upload handler for examples
-            app.post('/upload', function(req, res) {
+            app.post('/upload', (req, res) => {
                 // save uploaded file
                 var form = new formidable.IncomingForm();
                 form.uploadDir = 'uploads';
@@ -45,18 +45,18 @@ module.exports = {
 
                 console.log('saving uploaded file...');
 
-                form.on('fileBegin', function(name, file) {
+                form.on('fileBegin', (name, file) => {
                     // use original filename in this example
                     file.path = form.uploadDir + '/' + file.name;
                     console.log(colors.yellow('filename:', file.name));
                 });
 
-                form.on('end', function() {
+                form.on('end', () => {
                     console.log(colors.green('saved file.'));
                     console.log('');
                 });
 
-                form.parse(req, function(err, fields, files) {
+                form.parse(req, (err, fields, files) => {
                     res.writeHead(200, {'content-type': 'text/plain'});
                     res.write('received upload:\n\n');
                     res.end(util.inspect({fields: fields, files: files}));